Rename game.state and gui.list and disband mapPanel.minimap packages
