+++ src/main/java/org/openRealmOfStars/game/state/AmbientLightView.java
@@ -44,7 +44,6 @@ import org.openRealmOfStars.ambient.ServiceDiscovery;
 import org.openRealmOfStars.audio.soundeffect.SoundPlayer;
 import org.openRealmOfStars.game.GameCommands;
 import org.openRealmOfStars.game.config.ConfigFile;
-import org.openRealmOfStars.gui.ListRenderers.BridgeDeviceRenderer;
 import org.openRealmOfStars.gui.borders.SimpleBorder;
 import org.openRealmOfStars.gui.buttons.SpaceButton;
 import org.openRealmOfStars.gui.buttons.SpaceCheckBox;
@@ -53,6 +52,7 @@ import org.openRealmOfStars.gui.infopanel.EmptyInfoPanel;
 import org.openRealmOfStars.gui.infopanel.InfoPanel;
 import org.openRealmOfStars.gui.labels.InfoTextArea;
 import org.openRealmOfStars.gui.labels.SpaceLabel;
+import org.openRealmOfStars.gui.list.BridgeDeviceRenderer;
 import org.openRealmOfStars.gui.panels.BlackPanel;
 import org.openRealmOfStars.gui.panels.SpaceSliderPanel;
 import org.openRealmOfStars.gui.util.GuiStatics;
