+++ src/main/java/org/openRealmOfStars/game/state/DiplomacyView.java
@@ -40,16 +40,16 @@ import org.openRealmOfStars.audio.music.MusicPlayer;
 import org.openRealmOfStars.audio.soundeffect.SoundPlayer;
 import org.openRealmOfStars.game.Game;
 import org.openRealmOfStars.game.GameCommands;
-import org.openRealmOfStars.gui.ListRenderers.FleetListRenderer;
-import org.openRealmOfStars.gui.ListRenderers.PlanetListRenderer;
-import org.openRealmOfStars.gui.ListRenderers.SpeechLineRenderer;
-import org.openRealmOfStars.gui.ListRenderers.TechListRenderer;
 import org.openRealmOfStars.gui.buttons.SpaceButton;
 import org.openRealmOfStars.gui.buttons.SpaceCheckBox;
 import org.openRealmOfStars.gui.icons.Icons;
 import org.openRealmOfStars.gui.infopanel.InfoPanel;
 import org.openRealmOfStars.gui.labels.InfoTextArea;
 import org.openRealmOfStars.gui.labels.SpaceLabel;
+import org.openRealmOfStars.gui.list.FleetListRenderer;
+import org.openRealmOfStars.gui.list.PlanetListRenderer;
+import org.openRealmOfStars.gui.list.SpeechLineRenderer;
+import org.openRealmOfStars.gui.list.TechListRenderer;
 import org.openRealmOfStars.gui.panels.BlackPanel;
 import org.openRealmOfStars.gui.panels.ShipInteriorPanel;
 import org.openRealmOfStars.gui.panels.SpaceGreyPanel;
