+++ src/main/java/org/openRealmOfStars/gui/labels/PlanetInfoLabel.java
@@ -26,11 +26,11 @@ import javax.swing.BoxLayout;
 import javax.swing.JComboBox;
 
 import org.openRealmOfStars.game.GameCommands;
-import org.openRealmOfStars.gui.ListRenderers.ProductionListRenderer;
 import org.openRealmOfStars.gui.borders.SimpleBorder;
 import org.openRealmOfStars.gui.buttons.SpaceButton;
 import org.openRealmOfStars.gui.icons.Icons;
 import org.openRealmOfStars.gui.infopanel.EmptyInfoPanel;
+import org.openRealmOfStars.gui.list.ProductionListRenderer;
 import org.openRealmOfStars.gui.util.GuiStatics;
 import org.openRealmOfStars.mapTiles.Tile;
 import org.openRealmOfStars.mapTiles.Tiles;
