+++ src/main/java/org/openRealmOfStars/gui/list/ArtifactListRenderer.java
@@ -1,7 +1,7 @@
-package org.openRealmOfStars.gui.ListRenderers;
+package org.openRealmOfStars.gui.list;
 /*
  * Open Realm of Stars game project
- * Copyright (C) 2016 Tuomo Untinen
+ * Copyright (C) 2022 Tuomo Untinen
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -17,6 +17,7 @@ package org.openRealmOfStars.gui.ListRenderers;
  * along with this program; if not, see http://www.gnu.org/licenses/
  */
 
+import java.awt.Color;
 import java.awt.Component;
 
 import javax.swing.DefaultListCellRenderer;
@@ -25,14 +26,14 @@ import javax.swing.JList;
 import javax.swing.ListCellRenderer;
 
 import org.openRealmOfStars.gui.util.GuiStatics;
-import org.openRealmOfStars.player.ship.ShipHull;
+import org.openRealmOfStars.player.artifact.Artifact;
 
 /**
  *
- * Ship Hull list renderer
+ * Artifact list renderer
  *
  */
-public class ShipHullListRenderer implements ListCellRenderer<ShipHull> {
+public class ArtifactListRenderer implements ListCellRenderer<Artifact> {
 
   /**
    * Default list cell renderer
@@ -42,22 +43,19 @@ public class ShipHullListRenderer implements ListCellRenderer<ShipHull> {
 
   @Override
   public Component getListCellRendererComponent(
-      final JList<? extends ShipHull> list, final ShipHull value,
+      final JList<? extends Artifact> list, final Artifact value,
       final int index, final boolean isSelected, final boolean cellHasFocus) {
     JLabel renderer = (JLabel) defaultRenderer.getListCellRendererComponent(
         list, value, index, isSelected, cellHasFocus);
     renderer.setFont(GuiStatics.getFontCubellan());
-    if (value != null) {
-      renderer.setText(value.getName() + " - " + value.getSize().toString());
-    } else {
-      renderer.setText("No hull");
-    }
+    renderer.setIcon(value.getIcon().getAsIcon());
+    renderer.setText(value.getName());
     if (isSelected) {
-      renderer.setForeground(GuiStatics.getCoolSpaceColor());
-      renderer.setBackground(GuiStatics.getDeepSpaceColor());
+      renderer.setForeground(GuiStatics.getInfoTextColor());
+      renderer.setBackground(Color.BLACK);
     } else {
-      renderer.setForeground(GuiStatics.getCoolSpaceColorDark());
-      renderer.setBackground(GuiStatics.getDeepSpaceDarkColor());
+      renderer.setForeground(GuiStatics.getInfoTextColorDark());
+      renderer.setBackground(Color.BLACK);
     }
     return renderer;
   }
