+++ src/main/java/org/openRealmOfStars/gui/list/ShipComponentListRenderer.java
@@ -1,4 +1,4 @@
-package org.openRealmOfStars.gui.ListRenderers;
+package org.openRealmOfStars.gui.list;
 /*
  * Open Realm of Stars game project
  * Copyright (C) 2016 Tuomo Untinen
