+++ src/main/java/org/openRealmOfStars/gui/list/ShipStatRenderer.java
@@ -1,4 +1,4 @@
-package org.openRealmOfStars.gui.ListRenderers;
+package org.openRealmOfStars.gui.list;
 /*
  * Open Realm of Stars game project
  * Copyright (C) 2016 Tuomo Untinen
@@ -17,6 +17,7 @@ package org.openRealmOfStars.gui.ListRenderers;
  * along with this program; if not, see http://www.gnu.org/licenses/
  */
 
+import java.awt.Color;
 import java.awt.Component;
 
 import javax.swing.DefaultListCellRenderer;
@@ -25,14 +26,14 @@ import javax.swing.JList;
 import javax.swing.ListCellRenderer;
 
 import org.openRealmOfStars.gui.util.GuiStatics;
-import org.openRealmOfStars.starMap.planet.construction.Construction;
+import org.openRealmOfStars.player.ship.ShipStat;
 
 /**
  *
- * Production list renderer
+ * ShipStat list renderer
  *
  */
-public class ProductionListRenderer implements ListCellRenderer<Construction> {
+public class ShipStatRenderer implements ListCellRenderer<ShipStat> {
 
   /**
    * Default list cell renderer
@@ -42,19 +43,40 @@ public class ProductionListRenderer implements ListCellRenderer<Construction> {
 
   @Override
   public Component getListCellRendererComponent(
-      final JList<? extends Construction> list, final Construction value,
+      final JList<? extends ShipStat> list, final ShipStat value,
       final int index, final boolean isSelected, final boolean cellHasFocus) {
     JLabel renderer = (JLabel) defaultRenderer.getListCellRendererComponent(
         list, value, index, isSelected, cellHasFocus);
     renderer.setFont(GuiStatics.getFontCubellan());
-    renderer.setIcon(value.getIcon().getAsIcon());
-    if (isSelected) {
-      renderer.setForeground(GuiStatics.getCoolSpaceColor());
-      renderer.setBackground(GuiStatics.getDeepSpaceColor());
+    StringBuilder sb = new StringBuilder();
+    sb.append(value.getDesign().getName());
+    sb.append(" - ");
+    sb.append(value.getDesign().getHull().getSize());
+    if (value.isObsolete()) {
+      sb.append(" Obsoleted");
     } else {
-      renderer.setForeground(GuiStatics.getCoolSpaceColorDark());
-      renderer.setBackground(GuiStatics.getDeepSpaceDarkColor());
+      sb.append(" Cost/Metal: ");
+      sb.append(value.getDesign().getCost());
+      sb.append("/");
+      sb.append(value.getDesign().getMetalCost());
+    }
+    renderer.setText(sb.toString());
+    if (isSelected) {
+      if (value.isObsolete()) {
+        renderer.setForeground(GuiStatics.COLOR_GREY_TEXT);
+      } else {
+        renderer.setForeground(GuiStatics.getInfoTextColor());
+      }
+      renderer.setBackground(Color.BLACK);
+    } else {
+      if (value.isObsolete()) {
+        renderer.setForeground(GuiStatics.COLOR_GREY_TEXT_DARK);
+      } else {
+        renderer.setForeground(GuiStatics.getInfoTextColorDark());
+      }
+      renderer.setBackground(Color.BLACK);
     }
     return renderer;
   }
+
 }
