+++ src/main/java/org/openRealmOfStars/gui/list/TechListRenderer.java
@@ -1,4 +1,4 @@
-package org.openRealmOfStars.gui.ListRenderers;
+package org.openRealmOfStars.gui.list;
 /*
  * Open Realm of Stars game project
  * Copyright (C) 2016 Tuomo Untinen
@@ -17,6 +17,7 @@ package org.openRealmOfStars.gui.ListRenderers;
  * along with this program; if not, see http://www.gnu.org/licenses/
  */
 
+import java.awt.Color;
 import java.awt.Component;
 
 import javax.swing.DefaultListCellRenderer;
@@ -25,14 +26,14 @@ import javax.swing.JList;
 import javax.swing.ListCellRenderer;
 
 import org.openRealmOfStars.gui.util.GuiStatics;
-import org.openRealmOfStars.player.fleet.Fleet;
+import org.openRealmOfStars.player.tech.Tech;
 
 /**
  *
- * Fleet list renderer
+ * Tech list renderer
  *
  */
-public class FleetListRenderer implements ListCellRenderer<Fleet> {
+public class TechListRenderer implements ListCellRenderer<Tech> {
 
   /**
    * Default list cell renderer
@@ -42,24 +43,18 @@ public class FleetListRenderer implements ListCellRenderer<Fleet> {
 
   @Override
   public Component getListCellRendererComponent(
-      final JList<? extends Fleet> list, final Fleet value, final int index,
+      final JList<? extends Tech> list, final Tech value, final int index,
       final boolean isSelected, final boolean cellHasFocus) {
     JLabel renderer = (JLabel) defaultRenderer.getListCellRendererComponent(
         list, value, index, isSelected, cellHasFocus);
     renderer.setFont(GuiStatics.getFontCubellan());
-    if (value.getNumberOfShip() == 1) {
-      renderer
-          .setText(value.getName() + " - " + value.getFirstShip().getName());
-    } else {
-      renderer.setText(
-          value.getName() + " - " + value.getNumberOfShip() + " ships");
-    }
+    renderer.setIcon(value.getIcon().getAsIcon());
     if (isSelected) {
-      renderer.setForeground(GuiStatics.getCoolSpaceColor());
-      renderer.setBackground(GuiStatics.getDeepSpaceColor());
+      renderer.setForeground(GuiStatics.getInfoTextColor());
+      renderer.setBackground(Color.BLACK);
     } else {
-      renderer.setForeground(GuiStatics.getCoolSpaceColorDark());
-      renderer.setBackground(GuiStatics.getDeepSpaceDarkColor());
+      renderer.setForeground(GuiStatics.getInfoTextColorDark());
+      renderer.setBackground(Color.BLACK);
     }
     return renderer;
   }
