+++ src/main/java/org/openRealmOfStars/gui/list/TutorialTreeCellRenderer.java
@@ -1,7 +1,7 @@
-package org.openRealmOfStars.gui.ListRenderers;
+package org.openRealmOfStars.gui.list;
 /*
  * Open Realm of Stars game project
- * Copyright (C) 2022 Tuomo Untinen
+ * Copyright (C) 2019 Tuomo Untinen
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -17,6 +17,7 @@ package org.openRealmOfStars.gui.ListRenderers;
  * along with this program; if not, see http://www.gnu.org/licenses/
  */
 
+import java.awt.Color;
 import java.awt.Component;
 
 import javax.swing.JLabel;
@@ -25,16 +26,15 @@ import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.DefaultTreeCellRenderer;
 import javax.swing.tree.TreeCellRenderer;
 
+import org.openRealmOfStars.game.tutorial.HelpLine;
 import org.openRealmOfStars.gui.util.GuiStatics;
-import org.openRealmOfStars.player.leader.Leader;
-import org.openRealmOfStars.player.leader.LeaderUtility;
 
 /**
 *
 * Tutorial tree cell renderer.
 *
 */
-public class LeaderTreeCellRenderer implements TreeCellRenderer {
+public class TutorialTreeCellRenderer implements TreeCellRenderer {
 
   /**
    * Default Tree cell renderer
@@ -47,6 +47,13 @@ public class LeaderTreeCellRenderer implements TreeCellRenderer {
       final boolean leaf, final int row, final boolean hasFocus) {
     JLabel renderer = (JLabel) defaultRenderer.getTreeCellRendererComponent(
         tree, value, selected, expanded, leaf, row, hasFocus);
+    if (selected) {
+      renderer.setBackground(GuiStatics.getCoolSpaceColorDarker());
+      renderer.setForeground(GuiStatics.getInfoTextColor());
+    } else {
+      renderer.setBackground(Color.BLACK);
+      renderer.setForeground(GuiStatics.getInfoTextColorDark());
+    }
     if (value != null && value instanceof DefaultMutableTreeNode) {
       Object object = ((DefaultMutableTreeNode) value)
           .getUserObject();
@@ -54,12 +61,9 @@ public class LeaderTreeCellRenderer implements TreeCellRenderer {
         String str = (String) object;
         renderer.setText(str);
       }
-      if (object instanceof Leader) {
-        Leader leader = (Leader) object;
-        renderer.setText(leader.getCallName());
-        renderer.setFont(GuiStatics.getFontCubellan());
-        renderer.setIcon(
-            LeaderUtility.getIconBasedOnLeaderJob(leader).getAsIcon());
+      if (object instanceof HelpLine) {
+        HelpLine line = (HelpLine) object;
+        renderer.setText(line.getTitle());
       }
     }
     return renderer;
