+++ src/main/java/org/openRealmOfStars/gui/list/VotingListRenderer.java
@@ -1,7 +1,7 @@
-package org.openRealmOfStars.gui.ListRenderers;
+package org.openRealmOfStars.gui.list;
 /*
  * Open Realm of Stars game project
- * Copyright (C) 2016 Tuomo Untinen
+ * Copyright (C) 2022 Tuomo Untinen
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -17,6 +17,7 @@ package org.openRealmOfStars.gui.ListRenderers;
  * along with this program; if not, see http://www.gnu.org/licenses/
  */
 
+import java.awt.Color;
 import java.awt.Component;
 
 import javax.swing.DefaultListCellRenderer;
@@ -25,14 +26,15 @@ import javax.swing.JList;
 import javax.swing.ListCellRenderer;
 
 import org.openRealmOfStars.gui.util.GuiStatics;
-import org.openRealmOfStars.player.ship.Ship;
+import org.openRealmOfStars.starMap.StarMap;
+import org.openRealmOfStars.starMap.vote.Vote;
 
 /**
- *
- * Ship list renderer
- *
- */
-public class ShipListRenderer implements ListCellRenderer<Ship> {
+*
+* Voting list renderer
+*
+*/
+public class VotingListRenderer implements ListCellRenderer<Vote> {
 
   /**
    * Default list cell renderer
@@ -40,23 +42,34 @@ public class ShipListRenderer implements ListCellRenderer<Ship> {
   private DefaultListCellRenderer defaultRenderer
       = new DefaultListCellRenderer();
 
+  /**
+   * Starmap for getting enough information for voting.
+   */
+  private StarMap map;
+
+  /**
+   * Set Star map for renderer.
+   * @param starMap StarMap
+   */
+  public void setStarMap(final StarMap starMap) {
+    map = starMap;
+  }
   @Override
   public Component getListCellRendererComponent(
-      final JList<? extends Ship> list, final Ship value, final int index,
+      final JList<? extends Vote> list, final Vote value, final int index,
       final boolean isSelected, final boolean cellHasFocus) {
     JLabel renderer = (JLabel) defaultRenderer.getListCellRendererComponent(
         list, value, index, isSelected, cellHasFocus);
     renderer.setFont(GuiStatics.getFontCubellan());
-    renderer.setText(
-        value.getName() + " - " + value.getHull().getHullType().toString()
-        + " - " + value.getTotalMilitaryPower());
+    renderer.setText(value.getDescription(map));
     if (isSelected) {
-      renderer.setForeground(value.getDamageColor());
-      renderer.setBackground(GuiStatics.getDeepSpaceColor());
+      renderer.setForeground(GuiStatics.COLOR_GOLD);
+      renderer.setBackground(Color.BLACK);
     } else {
-      renderer.setForeground(value.getDamageColor());
-      renderer.setBackground(GuiStatics.getDeepSpaceDarkColor());
+      renderer.setForeground(GuiStatics.COLOR_GOLD_DARK);
+      renderer.setBackground(Color.BLACK);
     }
     return renderer;
   }
+
 }
